requirements.txt
Django==4.2
djangorestframework==3.14.0
pytest==7.3.1
